requirements.txt
minio
confluent-kafka
zenml
zenml[server]
kafka-python
jpype1
cv
rfdetr
supervision
ffmpeg-python